Update create_parallel_corridors functions, test successfully with v1 dec 2022 and v2 jan 2023
