+++ bus_service_increase/bus_service_utils/create_parallel_corridors.py
@@ -6,86 +6,63 @@ Highways Raw data: [SHN on Geoportal](https://opendata.arcgis.com/datasets/77f2d
 Processed data: run raw data through `create_parallel_corridors.py` 
 export to GCS, put in catalog.
 
-Transit Routes: Use `traffic_ops/export_shapefiles.py` that 
-creates `routes_assembled.parquet` in GCS, put in catalog.
-
-Instead of catalog[file_name].read(), put in GCS path
-because it's easier to import once these utility functions need to 
-be importable across directories.
+Transit Routes: Need a shapes level table with route info.
+Usually, this can be achieved by merging `trips` and `shapes`.
 """
 import geopandas as gpd
 import pandas as pd
 
-import shared_utils
-from bus_service_utils import utils
+from typing import Literal
+
+from shared_utils import geography_utils, utils
 
 DATA_PATH = "./data/"
 
 #--------------------------------------------------------#
-### State Highway Network
-#--------------------------------------------------------#
-def clean_highways():
-    HIGHWAY_URL = ("https://opendata.arcgis.com/datasets/"
-                   "77f2d7ba94e040a78bfbe36feb6279da_0.geojson")
-    gdf = gpd.read_file(HIGHWAY_URL)
-    
-    keep_cols = ['Route', 'County', 'District', 'RouteType',
-                 'Direction', 'geometry']
-    
-    gdf = gdf[keep_cols]
-    print(f"# rows before dissolve: {len(gdf)}")
-    
-    # See if we can dissolve further - use all cols except geometry
-    # Should we dissolve further and use even longer lines?
-    dissolve_cols = list(gdf.columns)
-    dissolve_cols.remove('geometry')
-    
-    gdf2 = gdf.dissolve(by=dissolve_cols).reset_index()
-    print(f"# rows after dissolve: {len(gdf2)}")
-    
-    # Export to GCS
-    shared_utils.utils.geoparquet_gcs_export(gdf2,
-                                         utils.GCS_FILE_PATH, 
-                                         "state_highway_network")
-    
-    
-#--------------------------------------------------------#
 ### Functions to create overlay dataset + further cleaning
 #--------------------------------------------------------#
-# Can this function be reworked to take a df?
-def process_transit_routes(alternate_df: 
-                           gpd.GeoDataFrame = None) -> gpd.GeoDataFrame:
-    if alternate_df is None:
-        df = gpd.read_parquet(
-            f"{utils.GCS_FILE_PATH}2022_Jan/shapes_processed.parquet")
-    else:
-        df = alternate_df
+def process_transit_routes(
+    df: gpd.GeoDataFrame, 
+    warehouse_version: Literal["v1", "v2"] = "v2"
+) -> gpd.GeoDataFrame:
+    """
+    At operator level, pick the route with the longest length 
+    to overlay with SHN.
+    At operator level, sum up how many unique routes there are.
+    """
+    
+    if warehouse_version == "v1":
+        operator_cols = ["calitp_itp_id"]
+        
+    elif warehouse_version == "v2":
+        operator_cols = ["feed_key"]
     
     ## Clean transit routes
-    df = df.to_crs(shared_utils.geography_utils.CA_StatePlane)
-
-    # Transit routes included some extra info about 
-    # route_long_name, route_short_name, agency_id
-    # Get it down to route_id instead of shape_id, pick longest shape
-    df2 = df.assign(route_length = df.geometry.length)
-    df3 = (df2.sort_values(["itp_id", "route_id", "route_length"], 
-                  ascending=[True, True, False])
-           .drop_duplicates(subset=["itp_id", "route_id"])
-          .reset_index(drop=True)
-         )
+    df = df.assign(
+        route_length = df.to_crs(
+            geography_utils.CA_StatePlane).geometry.length
+    ).to_crs(geography_utils.CA_StatePlane)
+    
+    # Get it down to route_id and pick longest shape
+    df2 = (df.sort_values(operator_cols + ["route_id", "route_length"], 
+                          ascending = [True, True, False])
+           .drop_duplicates(subset=operator_cols + ["route_id"])
+           .reset_index(drop=True)
+    )
     
     # Then count how many unique route_ids appear for operator (that's denominator)
-    keep_cols = [
-        "itp_id", "route_id", "total_routes",
-        "route_length", "geometry"
-    ]
+    route_cols = ["route_id", "total_routes", "route_length", "geometry"]
+    if warehouse_version == "v2":
+        keep_cols = operator_cols + ["name"] + route_cols
+    else:
+        keep_cols = operator_cols + route_cols
     
-    df4 = df3.assign(
-        total_routes = df3.groupby("itp_id")["route_id"].transform(
+    df3 = df2.assign(
+        total_routes = df2.groupby(operator_cols)["route_id"].transform(
             "nunique").astype("Int64")
     )[keep_cols]
     
-    return df4
+    return df3
 
 
 def prep_highway_directions_for_dissolve(
@@ -95,8 +72,9 @@ def prep_highway_directions_for_dissolve(
     Put in a list of group_cols, and aggregate highway segments with 
     the direction info up to the group_col level.
     '''
-    df = (gpd.read_parquet(f"{utils.GCS_FILE_PATH}state_highway_network.parquet")
-          .to_crs(shared_utils.geography_utils.CA_StatePlane))
+    df = (gpd.read_parquet("gs://calitp-analytics-data/data-analyses/"
+                           "state_highway_network.parquet")
+          .to_crs(geography_utils.CA_StatePlane))
     
     # Get dummies for direction
     # Can make data wide instead of long
@@ -110,11 +88,17 @@ def prep_highway_directions_for_dissolve(
         df[c] = df.groupby(group_cols)[c].transform('max').astype(int)
 
     return df
-    
-    
-def process_highways(buffer_feet: int = 
-                     shared_utils.geography_utils.FEET_PER_MI
-                    ) -> gpd.GeoDataFrame:
+
+
+def process_highways(
+    buffer_feet: int = geography_utils.FEET_PER_MI
+) -> gpd.GeoDataFrame:
+    """
+    For each highway, store what directions it runs in 
+    as dummy variables. This will allow us to dissolve 
+    the geometry and get fewer rows for highways
+    without losing direction info.
+    """
     
     group_cols = ["Route", "County", "District", "RouteType"]
     df = prep_highway_directions_for_dissolve(group_cols)
@@ -133,11 +117,12 @@ def process_highways(buffer_feet: int =
     df[direction_cols] = df[direction_cols].astype(int)
         
     return df
-    
-    
+
+
 def overlay_transit_to_highways(
-    hwy_buffer_feet: int = shared_utils.geography_utils.FEET_PER_MI,
-    alternate_df = None
+    hwy_buffer_feet: int = geography_utils.FEET_PER_MI,
+    transit_routes_df: gpd.GeoDataFrame = None, 
+    warehouse_version: Literal["v1", "v2"] = "v2"
 ) -> gpd.GeoDataFrame:
     """
     Function to find areas of intersection between
@@ -149,17 +134,17 @@ def overlay_transit_to_highways(
     
     # Can pass a different buffer zone to determine parallel corridors
     highways = process_highways(buffer_feet = hwy_buffer_feet)
-    transit_routes = process_transit_routes(alternate_df)
+    transit_routes = process_transit_routes(transit_routes_df, warehouse_version)
     
     # Overlay
     # Note: an overlay based on intersection changes the geometry column
     # The new geometry column will reflect that area of intersection
-    gdf = gpd.overlay(transit_routes, 
-                      highways, 
-                      how = "intersection", 
-                      keep_geom_type = False
-                     )  
-    
+    gdf = gpd.overlay(
+        transit_routes, 
+        highways, 
+        how = "intersection", 
+        keep_geom_type = False
+    )  
     
     # Using new geometry column, calculate what % that intersection 
     # is of the route and hwy
@@ -185,12 +170,20 @@ def overlay_transit_to_highways(
     
     # Fix geometry - don't want the overlay geometry, which is intersection
     # Want to use a transit route's line geometry
+    if warehouse_version == "v1":
+        operator_cols = ["calitp_itp_id"]
+        
+    elif warehouse_version == "v2":
+        operator_cols = ["feed_key", "name"]
+    
     gdf2 = pd.merge(
-        transit_routes[["itp_id", "route_id", "geometry"]].drop_duplicates(),
+        transit_routes[operator_cols + ["route_id", "geometry"]
+                      ].drop_duplicates(),
         gdf.drop(columns = "geometry"),
-        on = ["itp_id", "route_id"],
+        on = operator_cols + ["route_id"],
         how = "left",
-        # Allow 1:m merge because the same transit route can overlap with various highways
+        # Allow 1:m merge because the same transit route 
+        # can overlap with various highways
         validate = "1:m",
         indicator=True
     )
@@ -198,9 +191,11 @@ def overlay_transit_to_highways(
     return gdf2
 
 
-def parallel_or_intersecting(df: gpd.GeoDataFrame, 
-                             pct_route_threshold: float = 0.5, 
-                             pct_highway_threshold: float = 0.1) -> gpd.GeoDataFrame:
+def parallel_or_intersecting(
+    df: gpd.GeoDataFrame, 
+    pct_route_threshold: float = 0.5, 
+    pct_highway_threshold: float = 0.1
+) -> gpd.GeoDataFrame:
     
     # Play with various thresholds to decide how to designate parallel
     df = df.assign(
@@ -212,45 +207,55 @@ def parallel_or_intersecting(df: gpd.GeoDataFrame,
     )
     
     return df
- 
+
 
 # Use this in notebook
 # Can pass different parameters if buffer or thresholds need adjusting
-def make_analysis_data(hwy_buffer_feet=
-                       shared_utils.geography_utils.FEET_PER_MI,
-                       alternate_df = None,
-                       pct_route_threshold = 0.5,
-                       pct_highway_threshold = 0.1,
-                       DATA_PATH = "", FILE_NAME = "parallel_or_intersecting"
-                      ):
+def make_analysis_data(
+    hwy_buffer_feet:int = geography_utils.FEET_PER_MI,
+    transit_routes_df: gpd.GeoDataFrame = None,
+    pct_route_threshold: float = 0.5,
+    pct_highway_threshold: float = 0.1,
+    data_path: str = "", 
+    file_name: str = "parallel_or_intersecting",
+    warehouse_version: Literal["v1", "v2"] = "v2"
+):
     '''
     hwy_buffer_feet: int, number of feet to draw hwy buffers, defaults to 1 mile
-    alternate_df: None or pandas.DataFrame
-                    can input an alternate df for transit routes
-                    otherwise, use the `shapes_processed` from other analysis
+    transit_routes_df: a table, typically shapes, with route_id too
     pct_route_threshold: float, between 0-1
     pct_highway_threshold: float, between 0-1
-    DATA_PATH: str, file path for saving parallel transit routes dataset
-    FILE_NAME: str, file name for saving parallel transit routes dataset
+    data_path: str, file path for saving parallel transit routes dataset
+    file_name: str, file name for saving parallel transit routes dataset
     
     '''
     # Get overlay between highway and transit routes
     # Draw buffer around highways
-    gdf = overlay_transit_to_highways(hwy_buffer_feet, alternate_df)
+    gdf = overlay_transit_to_highways(
+        hwy_buffer_feet, transit_routes_df, warehouse_version)
 
     # Remove transit routes that do not overlap with highways
-    gdf = gdf[gdf._merge!="left_only"].drop(columns = "_merge").reset_index(drop=True)
+    gdf = (gdf[gdf._merge!="left_only"]
+           .drop(columns = "_merge")
+           .reset_index(drop=True)
+          )
     
     # Categorize whether route is parallel or intersecting based on threshold
-    gdf2 = parallel_or_intersecting(gdf, 
-                                    pct_route_threshold, 
-                                    pct_highway_threshold)
+    gdf2 = parallel_or_intersecting(
+        gdf, 
+        pct_route_threshold, 
+        pct_highway_threshold
+    )
     
-    if "gs://" in DATA_PATH:
-        shared_utils.utils.geoparquet_gcs_export(gdf2, DATA_PATH, FILE_NAME)
+    if "gs://" in data_path:
+        utils.geoparquet_gcs_export(
+            gdf2, 
+            data_path, 
+            file_name
+        )
     else:
-        gdf2.to_parquet(f"{DATA_PATH}{FILE_NAME}.parquet")
-    
+        gdf2.to_parquet(f"{data_path}{file_name}.parquet")
+        
     # For map, need highway to be 250 ft buffer
     #highways = process_highways(buffer_feet=250)
     #highways.to_parquet(f"{DATA_PATH}highways.parquet")
